Moved InvocationHandler into InterfaceSpecimen
